refactor: 모든 탭 내 섹션 모두 별도의 컴포넌트로 분리
